src/containers/client/requirements.txt
requests
lumigo_opentelemetry
opentelemetry-exporter-jaeger